@@ -38,7 +38,7 @@ class GoodInfluxUdp extends Stream.Writable {
         // Standard users
         this.once('finish', () => {
 
-            this._sendMessages();
+            this._sendMessages(() => { /* no-op */ });
         });
         this.config = Hoek.applyToDefaults(internals.defaults, config);
         this.schemas = schemas;
@@ -67,7 +67,7 @@ class GoodInfluxUdp extends Stream.Writable {
             lineData = lineData.concat(lines);
         });
 
-        const payload = new Buffer(lineData.join('\n'));
+        const payload = Buffer.from(lineData.join('\n'));
 
         // Prevent this from user tampering
         this._udpClient.send(payload, 0, payload.length, this._endpoint.port, this._endpoint.hostname, callback);
