@@ -38,7 +38,7 @@ class GoodInfluxHttp extends Stream.Writable {
         // Standard users
         this.once('finish', () => {
 
-            this._sendMessages();
+            this._sendMessages(() => { /* no-op */ });
         });
         this.config = Hoek.applyToDefaults(internals.defaults, config);
         this.schemas = schemas;
@@ -79,7 +79,7 @@ class GoodInfluxHttp extends Stream.Writable {
         Hoek.merge(wreckOptions, this._settings.wreck, { nullOverride: false });
 
         // Prevent this from user tampering
-        if (wreckOptions.headers){
+        if (wreckOptions.headers) {
             wreckOptions.headers['content-type'] = 'text/plain';
         }
         else {
@@ -87,6 +87,8 @@ class GoodInfluxHttp extends Stream.Writable {
         }
         Wreck.request('post', this._endpoint, wreckOptions).then((req) => {
             return callback(null, req);
+        }).catch((err) => {
+            return callback(err);
         });
     }
 }
